+++ lib/client/doc.js
@@ -504,7 +504,7 @@ Doc.prototype.flush = function() {
 
 // Helper function to set op to contain a no-op.
 function setNoOp(op) {
-  delete op.op;
+  op.op = null;
   delete op.create;
   delete op.del;
 }
@@ -528,7 +528,7 @@ function transformX(client, server) {
   }
 
   // Ignore no-op coming from server
-  if (!('op' in server)) return;
+  if (!server.op) return;
 
   // I believe that this should not occur, but check just in case
   if (client.create) {
